Modal de cuentas

+++ src/services/CuentaInvocadorService.ts
@@ -1,6 +1,10 @@
 import { Repository } from 'typeorm';
+import axios from 'axios';
+
 import { CuentaInvocador } from '@/entities/CuentaInvocadorEntity';
 import AppDataSource from '@/config/ormconfig';
+import { AuthService } from '@/services/misc/AuthService';
+import { BadRequestError, NotFoundError } from '@/middlewares/appError';
 
 export class CuentaInvocadorService {
     private readonly cuentaInvocadorRepository: Repository<CuentaInvocador>;
@@ -13,12 +17,32 @@ export class CuentaInvocadorService {
     async createCuentaInvocador(
         data: Partial<CuentaInvocador>,
     ): Promise<CuentaInvocador> {
+        await this.validateUniqueNombreTag(
+            data.nombreInvocador!,
+            data.tagInvocador!,
+        );
+
+        // Validar el invocador en la API de Riot y obtener el puuid
+        const { puuid } = await this.validateCuentaInRiotAPI(
+            data.nombreInvocador!,
+            data.tagInvocador!,
+        );
+
+        // Llamar al nuevo endpoint para obtener el id
+        const riotId = await this.getRiotIdByPuuid(puuid);
+
+        data.puuid = puuid;
+        data.cuentaId = riotId;
+
         const cuentaInvocador = this.cuentaInvocadorRepository.create(data);
         return await this.cuentaInvocadorRepository.save(cuentaInvocador);
     }
 
     async getCuentaInvocadorById(id: number): Promise<CuentaInvocador | null> {
-        return await this.cuentaInvocadorRepository.findOneBy({ id });
+        return await this.cuentaInvocadorRepository.findOne({
+            where: { id },
+            relations: ['usuario'],
+        });
     }
 
     async updateCuentaInvocador(
@@ -29,11 +53,106 @@ export class CuentaInvocadorService {
         return this.getCuentaInvocadorById(id);
     }
 
-    async deleteCuentaInvocador(id: number): Promise<void> {
+    async deleteCuentaInvocador(id: number, token: string): Promise<void> {
+        const decodedToken = AuthService.verifyToken(token);
+        if (typeof decodedToken === 'string') {
+            throw new Error('Invalid token');
+        }
+
+        const userId = decodedToken.id;
+        const cuentaInvocador = await this.getCuentaInvocadorById(id);
+
+        if (!cuentaInvocador) {
+            throw new NotFoundError('Esta cuenta no existe');
+        }
+
+        if (cuentaInvocador.usuario.id !== userId) {
+            throw new BadRequestError('Esta no es tu cuenta');
+        }
+
         await this.cuentaInvocadorRepository.delete(id);
     }
 
     async getAllCuentaInvocadores(): Promise<CuentaInvocador[]> {
         return await this.cuentaInvocadorRepository.find();
+    }
+
+    async updateNombreYTagInvocador(puuid: string): Promise<CuentaInvocador | null> {
+        const riotApiKey = process.env.RIOT_API_KEY;
+        const apiUrl = `https://europe.api.riotgames.com/riot/account/v1/accounts/by-puuid/${puuid}?api_key=${riotApiKey}`;
+
+        try {
+            const response = await axios.get(apiUrl);
+            const { gameName, tagLine } = response.data;
+
+            const cuentaInvocador = await this.cuentaInvocadorRepository.findOne({
+                where: { puuid },
+            });
+
+            if (!cuentaInvocador) {
+                throw new NotFoundError('Cuenta de invocador no encontrada.');
+            }
+
+            cuentaInvocador.nombreInvocador = gameName;
+            cuentaInvocador.tagInvocador = tagLine;
+
+            return await this.cuentaInvocadorRepository.save(cuentaInvocador);
+        } catch (error: any) {
+            if (error.response && error.response.status === 404) {
+                throw new BadRequestError('La cuenta de invocador no existe en la API de Riot.');
+            }
+            throw new BadRequestError('Error al actualizar el nombre y tag del invocador.');
+        }
+    }
+
+    private async validateUniqueNombreTag(
+        nombreInvocador: string,
+        tagInvocador: string,
+    ): Promise<void> {
+        const existingCuenta = await this.cuentaInvocadorRepository.findOne({
+            where: { nombreInvocador, tagInvocador },
+        });
+
+        if (existingCuenta) {
+            throw new BadRequestError(
+                'Ya existe una cuenta con este nombre y tag.',
+            );
+        }
+    }
+
+    private async validateCuentaInRiotAPI(
+        nombreInvocador: string,
+        tagInvocador: string,
+    ): Promise<{ puuid: string }> {
+        const riotApiKey = process.env.RIOT_API_KEY;
+        const apiUrl = `https://europe.api.riotgames.com/riot/account/v1/accounts/by-riot-id/${encodeURIComponent(
+            nombreInvocador,
+        )}/${encodeURIComponent(tagInvocador)}`;
+
+        try {
+            const response = await axios.get(apiUrl, {
+                headers: {
+                    'X-Riot-Token': riotApiKey,
+                },
+            });
+
+            const { puuid } = response.data;
+            return { puuid };
+        } catch (error) {
+            throw new BadRequestError('El invocador introducido no existe.');
+        }
+    }
+
+    private async getRiotIdByPuuid(puuid: string): Promise<string> {
+        const riotApiKey = process.env.RIOT_API_KEY;
+        const apiUrl = `https://euw1.api.riotgames.com/lol/summoner/v4/summoners/by-puuid/${puuid}?api_key=${riotApiKey}`;
+
+        try {
+            const response = await axios.get(apiUrl);
+            const { id } = response.data;
+            return id;
+        } catch (error) {
+            throw new BadRequestError('Error al obtener el id de Riot.');
+        }
     }
 }
